requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.3
Django==5.0.2
django-oauth-toolkit==2.3.0
django-oidc-provider==0.8.2
djangorestframework==3.14.0
future==0.18.3
idna==3.6
jwcrypto==1.5.4
kombu==5.3.5
oauthlib==3.2.2
prompt-toolkit==3.0.43
pycparser==2.21
pycryptodomex==3.20.0
pyjwkest==1.4.2
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
